archive/requirements.txt
langgraph>=1.0.0
langchain>=1.0.0
langchain-google-genai>=1.0.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
aiofiles>=23.2.1
python-magic>=0.4.27
python-docx>=1.1.2
PyPDF2>=3.0.1
openpyxl>=3.1.5
pandas>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.31.0
